requirements.txt
numpy==2.2.6
opencv_python==4.12.0.88
Pillow==12.0.0
